refactor: reorganizar projeto e ajustar importações
- Mover projeto para o diretório backend
- Ajustar caminhos de importação nos arquivos de código
- Atualizar módulo no go.mod
