+++ backend/internal/api/api.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"sync"
 
-	"github.com/Kianelc/ama-room.git/internal/store/pgstore"
+	"github.com/Kianelc/ama-room/backend/internal/store/pgstore"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
